refactor(KONFLUX-10143): break import cycle between journey and logging
The `logging.Measure()` function requires access to the `ThreadIndex`
field of the `MainContext` struct for more detailed logging.

Attempting to add a type assertion for `*journey.MainContext` within the
`logging` package created an import cycle, as the `journey` package
already depends on `logging`.

To resolve this, a new `pkg/types` package has been introduced to hold
shared context structs (`MainContext`, `PerApplicationContext`,
`PerComponentContext`). Both the `journey` and `logging` packages now
import `types`, breaking the circular dependency.

All references to these context structs have been updated throughout the
codebase to use the new `types` package.

Generated-By: Gemini
