+++ tests/load-tests/pkg/journey/handle_applications.go
@@ -4,6 +4,7 @@ import "fmt"
 import "time"
 
 import logging "github.com/konflux-ci/e2e-tests/tests/load-tests/pkg/logging"
+import types "github.com/konflux-ci/e2e-tests/tests/load-tests/pkg/types"
 
 import framework "github.com/konflux-ci/e2e-tests/pkg/framework"
 import utils "github.com/konflux-ci/e2e-tests/pkg/utils"
@@ -34,7 +35,7 @@ func validateApplication(f *framework.Framework, name, namespace string) error {
 	return err
 }
 
-func HandleApplication(ctx *PerApplicationContext) error {
+func HandleApplication(ctx *types.PerApplicationContext) error {
 	var err error
 
 	logging.Logger.Debug("Creating application %s in namespace %s", ctx.ApplicationName, ctx.ParentContext.Namespace)
