+++ tests/load-tests/pkg/types/types.go
@@ -1,0 +1,48 @@
+package types
+
+import "sync"
+import "time"
+
+import framework "github.com/konflux-ci/e2e-tests/pkg/framework"
+import loadtestutils "github.com/konflux-ci/e2e-tests/tests/load-tests/pkg/loadtestutils"
+import options "github.com/konflux-ci/e2e-tests/tests/load-tests/pkg/options"
+
+// Struct to hold user journey thread data
+type MainContext struct {
+	ThreadsWG              *sync.WaitGroup
+	ThreadIndex            int
+	StartupPause           time.Duration
+	JourneyRepeatsCounter  int
+	Opts                   *options.Opts
+	StageUsers             *[]loadtestutils.User
+	Framework              *framework.Framework
+	Username               string
+	Namespace              string
+	ComponentRepoUrl       string // overrides same value from Opts, needed when templating repos
+	PerApplicationContexts []*PerApplicationContext
+}
+
+// Struct to hold data for thread to process each application
+type PerApplicationContext struct {
+	PerApplicationWG            *sync.WaitGroup
+	ApplicationIndex            int
+	StartupPause                time.Duration
+	Framework                   *framework.Framework
+	ParentContext               *MainContext
+	ApplicationName             string
+	IntegrationTestScenarioName string
+	PerComponentContexts        []*PerComponentContext
+}
+
+// Struct to hold data for thread to process each component
+type PerComponentContext struct {
+	PerComponentWG     *sync.WaitGroup
+	ComponentIndex     int
+	StartupPause       time.Duration
+	Framework          *framework.Framework
+	ParentContext      *PerApplicationContext
+	ComponentName      string
+	SnapshotName       string
+	MergeRequestNumber int
+	ReleaseName        string
+}
